방 둘러보기 기능 ViewSwitcher에 통합
- FPV 버튼 삭제
- 방 둘러보기 기능 ViewSwitcher에 통합
