+++ Assets/Scripts/CameraController/TopViewCameraController.cs
@@ -9,30 +9,12 @@ public class TopViewCameraController : MonoBehaviour
     public float minZoomDistance = 5.0f; // 최소 줌 거리
     public float maxZoomDistance = 100.0f; // 최대 줌 거리
 
-    public GameObject viewSwitcher;
     private Vector3 dragOrigin; // 드래그 시작점
     private Camera currentCamera; // 현재 카메라에 대한 참조
-    public Camera firstPersonViewCamera;
-    public TMP_Dropdown roomsDropdown;
 
     void Start()
     {
         currentCamera = GetComponent<Camera>(); // 현재 오브젝트에 부착된 카메라 컴포넌트를 가져옵니다.
-        roomsDropdown.value = 0; // Hallway 선택
-
-        // 드랍다운 값 변경 이벤트에 MoveToRoom 함수 연결
-        roomsDropdown.onValueChanged.AddListener(delegate {
-            DropdownValueChanged(roomsDropdown);
-        });
-    }
-
-    void DropdownValueChanged(TMP_Dropdown change)
-    {
-        // 선택된 방 이름 가져오기
-        string selectedRoom = change.options[change.value].text;
-        // 선택된 방으로 이동
-        viewSwitcher.SwitchToFirstPersonView();
-        firstPersonViewCamera.MoveToRoom(selectedRoom);
     }
 
     void Update()
